+++ vscode/src/chat/InlineChatViewProvider.ts
