+++ vscode/src/chat/MessageProvider.ts
@@ -15,19 +15,19 @@ import { IntentDetector } from '@sourcegraph/cody-shared/src/intent-detector'
 import { ANSWER_TOKENS, DEFAULT_MAX_TOKENS } from '@sourcegraph/cody-shared/src/prompt/constants'
 import { Message } from '@sourcegraph/cody-shared/src/sourcegraph-api'
 
-import { VSCodeEditor } from '../../editor/vscode-editor'
-import { logEvent } from '../../event-logger'
-import { debug } from '../../log'
-import { FixupTask } from '../../non-stop/FixupTask'
-import { IdleRecipeRunner } from '../../non-stop/roles'
-import { AuthProvider } from '../../services/AuthProvider'
-import { LocalStorage } from '../../services/LocalStorageProvider'
-import { TestSupport } from '../../test-support'
-import { fastFilesExist } from '../fastFileFinder'
-import { defaultAuthStatus } from '../protocol'
-import { getRecipe } from '../recipes'
+import { VSCodeEditor } from '../editor/vscode-editor'
+import { logEvent } from '../event-logger'
+import { debug } from '../log'
+import { FixupTask } from '../non-stop/FixupTask'
+import { IdleRecipeRunner } from '../non-stop/roles'
+import { AuthProvider } from '../services/AuthProvider'
+import { LocalStorage } from '../services/LocalStorageProvider'
+import { TestSupport } from '../test-support'
 
 import { ContextProvider } from './ContextProvider'
+import { fastFilesExist } from './fastFileFinder'
+import { defaultAuthStatus } from './protocol'
+import { getRecipe } from './recipes'
 
 export type Config = Pick<
     ConfigurationWithAccessToken,
