+++ vscode/src/main.ts
@@ -5,11 +5,11 @@ import { CodebaseContext } from '@sourcegraph/cody-shared/src/codebase-context'
 import { Configuration, ConfigurationWithAccessToken } from '@sourcegraph/cody-shared/src/configuration'
 import { SourcegraphNodeCompletionsClient } from '@sourcegraph/cody-shared/src/sourcegraph-api/completions/nodeClient'
 
+import { ChatViewProvider } from './chat/ChatViewProvider'
+import { ContextProvider } from './chat/ContextProvider'
+import { InlineChatViewProvider } from './chat/InlineChatViewProvider'
+import { MessageProviderOptions } from './chat/MessageProvider'
 import { CODY_FEEDBACK_URL } from './chat/protocol'
-import { ChatViewProvider } from './chat/v2/ChatViewProvider'
-import { ContextProvider } from './chat/v2/ContextProvider'
-import { InlineChatViewProvider } from './chat/v2/InlineChatViewProvider'
-import { MessageProviderOptions } from './chat/v2/MessageProvider'
 import { CodyCompletionItemProvider } from './completions'
 import { CompletionsDocumentProvider } from './completions/docprovider'
 import { History } from './completions/history'
